move code from the App to a separate component
