+++ src/components/RenderMeta.tsx
@@ -1,0 +1,19 @@
+function renderMeta(response: any) {
+  const usage = response?.data?.usage
+  const model = response?.data?.model
+  if (!usage) return ''
+  return (
+    <>
+      {', '}
+      <span
+        title={`${usage.prompt_tokens} prompt, ${usage.completion_tokens} completion`}
+      >
+        {usage.total_tokens} tokens
+      </span>
+      {', '}
+      {model}
+    </>
+  )
+}
+
+export default renderMeta
