+++ src/keys/StorageKey.ts
@@ -1,0 +1,5 @@
+function storageKey(nodeId: string) {
+  return `threadgpt/${nodeId}`
+}
+
+export default storageKey
