+++ src/types/Message.ts
@@ -1,0 +1,11 @@
+type Role = 'system' | 'user' | 'assistant'
+
+type Message = {
+  role: Role
+  content: string
+}
+
+export type {
+    Role,
+    Message
+}
